fix: reduce size of line and column number fields in tokenizer for storage efficiency
Also converted token class from int (64-bit) to int8 value size. May not affect storage
at this point (packing optimizations up to the Go compiler) but if the token grows in
complexity due to adding in line/column data, this will help.

+++ tokenizer/line.go
@@ -12,10 +12,10 @@ func (t *Tokenizer) SetLineNumber(line int) error {
 		return nil
 	}
 
-	currentLine := t.Line[t.TokenP]
+	currentLine := int(t.Line[t.TokenP])
 
-	offset := line - currentLine - 1
-	if offset > len(t.Line) {
+	offset := int32(line - currentLine - 1)
+	if offset > int32(len(t.Line)) {
 		return nil // nothing to do.
 	}
 
@@ -134,7 +134,7 @@ func (t *Tokenizer) Remainder() string {
 	p := t.Pos[t.TokenP] - 1
 	s := t.GetSource()
 
-	if p < 0 || p >= len(s) {
+	if p < 0 || p >= int32(len(s)) {
 		return ""
 	}
